# perf: read package.json only once in a run