## Summary

I tried the latest preview build on my private repository and observed slowdown from 3s to 24s. After profiling a run, I noticed that `ProjectLayout::get_node_manifest_for_path` is called for each rule, taking about 10 ms each.

Reading `package.json` was added for determining rule domains to apply automatically, but we don't need to read it for each rule. I changed `LintVisitor::new` to simply get the read manifest, and `AnalyzerVisitorBuilder::finish` to read the manifest only once.

I re-run Biome on the same repository after the changes, and it takes only 4s for entire run now.

<img width="1203" alt="image" src="https://github.com/user-attachments/assets/31829601-26fe-45f0-94c6-2e67988899a0" />


## Test Plan

Existing tests should pass. Plus, manually tested.