@siketyan 2025-03-17
@arendjr FYI, this is the breakdown of a `ProjectLayout::get_node_manifest_for_path` invocation. `Utf8Path::strip_prefix` takes 8 ms 🤯

<img width="1147" alt="image" src="https://github.com/user-attachments/assets/494812d9-7d4e-490b-ad00-86bfdc4fdfd9" />