@arendjr 2025-03-17
Oh, maybe it helps to use `starts_with()` instead. We also seem to call `strip_prefix()` in order to check if there _is_ a prefix.